requirements.txt
pymilvus
FlagEmbedding
Flask
python-docx